Move search tests into search.rs
